+++ Products/EasyNewsletter/setuphandlers.py
@@ -34,6 +34,7 @@ def add_catalog_indexes(context, logger=None):
     indexes = catalog.indexes()
     # Specify the indexes you want, with ('index_name', 'index_type')
     wanted = (('fullname', 'FieldIndex'),
+              ('nl_language', 'FieldIndex'),
               ('email', 'FieldIndex'),
               ('organization', 'FieldIndex'),
               )